updating campground authorization

+++ routes/campgrounds.js
@@ -66,11 +66,17 @@ router.get(
   '/:id/edit',
   isLoggedIn,
   asyncErrorWrapper(async (req, res) => {
-    const campground = await Campground.findById(req.params.id);
+    const { id } = req.params
+    const campground = await Campground.findById(id);
 
     if (!campground) {
       req.flash('error', 'Cannot find that campground');
       res.redirect('/campgrounds');
+    }
+
+    if (!campground.author.equals(req.user._id)) {
+      req.flash('error', 'You do not have permission to edit this campground');
+      return res.redirect(`/campgrounds/${id}`);
     }
 
     res.render('campgrounds/edit', { campground });
@@ -83,11 +89,16 @@ router.put(
   validateCampground,
   asyncErrorWrapper(async (req, res) => {
     const { id } = req.params;
-    const campground = await Campground.findByIdAndUpdate(id, {
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+      req.flash('error', 'You do not have permission to edit this campground');
+      return res.redirect(`/campgrounds/${id}`);
+    }
+    const camp = await Campground.findByIdAndUpdate(id, {
       ...req.body.campground,
     });
-    req.flash('success', `${campground.title} was successfully updated`);
-    res.redirect(`/campgrounds/${campground._id}`);
+    req.flash('success', `${camp.title} was successfully updated`);
+    res.redirect(`/campgrounds/${camp._id}`);
   })
 );
 
@@ -96,8 +107,13 @@ router.delete(
   isLoggedIn,
   asyncErrorWrapper(async (req, res) => {
     const { id } = req.params;
-    const campground = await Campground.findByIdAndDelete(id);
-    req.flash('success', `${campground.title} was successfully deleted`);
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+      req.flash('error', 'You do not have permission to delete this campground');
+      return res.redirect(`/campgrounds/${id}`);
+    }
+    const camp = await Campground.findByIdAndDelete(id);
+    req.flash('success', `${camp.title} was successfully deleted`);
     res.redirect('/campgrounds');
   })
 );
